+++ changelog.md
@@ -1,6 +1,7 @@
 * Agregado modulo de detector de idiomas.
+* Agregado Activar el intercambio automático si el origen detectado coincide con el destino (experimental)
+* Agregada ayuda en los diálogos al pulsar Ctrl+H.
 * Solucionado problema con la verbalización de algunos mensajes.
 * Cambiada forma de comprobar si hay internet.
-* Agregado retardo para iniciar el complemento si no se encuentra conexión a internet.
 *Corrección de errores
 - SHA256: